transition to ember-cli-build.js
