+++ ember-cli-build.js
@@ -1,17 +1,56 @@
 /* global require, module */
 var EmberApp = require('ember-cli/lib/broccoli/ember-addon');
+const Handlebars = require('handlebars'),
+  mergeTrees = require('broccoli-merge-trees'),
+  broccoliHandlebars = require('broccoli-handlebars'),
+  merge = require('lodash/object/merge');
+const sourceTrees = [];
 
 module.exports = function(defaults) {
   var app = new EmberApp(defaults, {
-    // Add options here
+    jscsOptions: {
+      enabled: true,
+      testGenerator: function(relativePath, errors) {
+        if (errors) {
+          errors = "\\n" + this.escapeErrorString(errors);
+        } else {
+          errors = "";
+        }
+
+        return "describe('JSCS - " + relativePath + "', function(){\n" +
+          "it('should pass jscs', function() { \n" +
+          "  expect(" + !errors + ", '" + relativePath + " should pass jscs." + errors + "').to.be.ok; \n" +
+          "})});\n";
+      }
+    }
   });
 
-  /*
-    This build file specifes the options for the dummy test app of this
-    addon, located in `/tests/dummy`
-    This build file does *not* influence how the addon or the app using it
-    behave. You most likely want to be modifying `./index.js` or app's build file
-  */
+  app.import('bower_components/bootstrap/dist/css/bootstrap.css');
 
-  return app.toTree();
+  sourceTrees.push(app.toTree());
+
+  // Docs
+  const cmdOpts = process.argv.slice(3);
+
+  if (cmdOpts.indexOf('--docs') !== -1) {
+    const docs = broccoliHandlebars('docs/theme', ['index.hbs'], {
+      handlebars: Handlebars,
+      helpers: require('./docs/theme/helpers'),
+      partials: 'docs/theme/partials',
+      context: function() {
+        return merge(
+          {},
+          require('./docs/config.json'),
+          require('./tmp/docs/data.json')
+        )
+      },
+      destFile: function (filename) {
+        return 'docs/' + filename.replace(/(hbs|handlebars)$/, 'html');
+      }
+    });
+
+    sourceTrees.push(docs);
+  }
+
+  return mergeTrees(sourceTrees);
 };
